console: Reduce size of background ID variable
The background functions only return values 0 to 7, so we can save some
space this way.

+++ include/nds/arm9/console.h
@@ -235,7 +235,7 @@ typedef struct PrintConsole
     u8 fontPalIndex;
 
     /// Background ID. Initialized by consoleInit().
-    int bgId;
+    s8 bgId;
 
     /// Current X location of the cursor. Initialized by consoleInit().
     s16 cursorX;
